hybrid Service implemented
